requirements.txt
python-telegram-bot[job-queue, httpx]==21.4
firebase-admin
nest_asyncio
matplotlib
googlemaps